# backend/public/swagger/swagger.yaml
openapi: 3.0.0
info:
  title: "O'Secours"
  description: 'API pour gérer des ressources'
  version: '1.0'
servers:
  -
    url: 'http://localhost:8000/api'
    description: 'API Server'
paths:
  /animals/all:
    get:
      tags:
        - Animals
      operationId: c7546f9cd96a5e416aca6b2c734067a4
      responses:
        '200':
          description: 'get list of all animals'
      security:
        -
          bearerAuth: []
  /animals:
    post:
      tags:
        - Animals
      summary: 'register new animal'
      description: 'Store animal informations with identification associated'
      operationId: createAnimal
      requestBody:
        description: 'Animal fields'
        required: true
        content:
          application/json:
            schema:
              required:
                - specie_id
              properties:
                name:
                  type: string
                  maxLength: 100
                  example: Pupuce
                description:
                  type: string
                  example: 'Really cute'
                birth_date:
                  type: string
                  format: '1995-04-01'
                  example: Pupuce
                cats_friendly:
                  type: boolean
                  example: true
                dogs_friendly:
                  type: boolean
                  example: true
                children_friendly:
                  type: boolean
                  example: true
                age:
                  type: integer
                  example: 4
                behavioral_comment:
                  type: string
                  example: 'Love cuddle'
                sterilized:
                  type: boolean
                  example: true
                deceased:
                  type: boolean
                  example: false
                specie_id:
                  type: integer
                  example: 1
                gender_id:
                  type: integer
                  example: 2
                color_id:
                  type: integer
                  example: 3
                sizerange_id:
                  type: integer
                  example: 4
                agerange_id:
                  type: integer
                  example: 2
                breed_id:
                  type: integer
                  example: 2
                number:
                  type: string
                  maxLength: 15
                  example: 152A5P6
              type: object
      responses:
        '200':
          description: 'Animal created successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Animal'
        '400':
          description: 'Bad request'
        '401':
          description: Unauthorized
  '/animals/{id}':
    get:
      tags:
        - Animals
      summary: 'Get specific animal details'
      description: 'Return informations for an animal with specific id'
      operationId: getAnimalById
      parameters:
        -
          name: id
          in: path
          description: 'animal Id to fetch'
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Animal fetched successfully'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Animal'
        '404':
          description: 'Animal not found'
        '500':
          description: 'Server Error'
    put:
      tags:
        - Animals
      summary: 'Update an animal'
      operationId: updateAnimalById
      parameters:
        -
          name: id
          in: path
          description: 'animal Id to update'
          required: true
          schema:
            type: string
      requestBody:
        description: 'Animal fields'
        required: true
        content:
          application/json:
            schema:
              properties:
                name:
                  type: string
                  maxLength: 100
                  example: Pupuce
                description:
                  type: string
                  example: 'Really cute'
                birth_date:
                  type: string
                  format: '1995-04-01'
                  example: Pupuce
                cats_friendly:
                  type: boolean
                  example: true
                dogs_friendly:
                  type: boolean
                  example: true
                children_friendly:
                  type: boolean
                  example: true
                age:
                  type: integer
                  example: 4
                behavioral_comment:
                  type: string
                  example: 'Love cuddle'
                sterilized:
                  type: boolean
                  example: true
                deceased:
                  type: boolean
                  example: false
                specie_id:
                  type: integer
                  example: 1
                gender_id:
                  type: integer
                  example: 2
                color_id:
                  type: integer
                  example: 3
                sizerange_id:
                  type: integer
                  example: 4
                agerange_id:
                  type: integer
                  example: 2
                breed_id:
                  type: integer
                  example: 2
                number:
                  type: string
                  maxLength: 15
                  example: 152A5P6
              type: object
      responses:
        '200':
          description: 'Animal created successfully'
        '400':
          description: 'Bad request'
        '401':
          description: Unauthorized
    delete:
      tags:
        - Animals
      summary: 'Delete specific animal (soft delete)'
      description: 'Make a soft delete for an animal with specific id'
      operationId: deleteAnimalById
      parameters:
        -
          name: id
          in: path
          description: "ID de l'animal à récupérer"
          required: true
          schema:
            type: string
      responses:
        '200':
          description: 'Animal récupéré avec succès'
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Animal'
        '404':
          description: 'Animal non trouvé'
        '500':
          description: 'Erreur interne'
...
components:
  schemas:
...
    Animal:
      title: Animal
      description: 'Animal model representing the detailed information of an animal in the system'
      properties:
        id:
          description: 'Unique identifier for the Animal'
          type: integer
          format: int64
          example: 1
        name:
          description: 'Name of the animal'
          type: string
          example: Bella
        description:
          description: 'Description of the animal'
          type: string
          example: 'A very playful young dog'
        birth_date:
          description: 'Birth date of the animal'
          type: string
          format: date
          example: '2017-04-21'
        cats_friendly:
          description: 'Indicates if the animal is friendly towards cats'
          type: boolean
          example: true
        dogs_friendly:
          description: 'Indicates if the animal is friendly towards dogs'
          type: boolean
          example: false
        children_friendly:
          description: 'Indicates if the animal is friendly towards children'
          type: boolean
          example: true
        age:
          description: 'Age of the animal'
          type: integer
          format: int32
          example: 3
        behavioral_comment:
          description: 'Behavioral comments about the animal'
          type: string
          example: 'Needs a lot of exercise'
        sterilized:
          description: 'Indicates if the animal is sterilized'
          type: boolean
          example: true
        deceased:
          description: 'Indicates if the animal is deceased'
          type: boolean
          example: false
        specie_id:
          description: 'Foreign key identifier for the species to which the animal belongs'
          type: integer
          format: int64
          example: 1
        breed_id:
          description: 'Foreign key identifier for the breed to which the animal belongs'
          type: integer
          format: int64
          example: 2
        gender_id:
          description: 'Foreign key identifier for the gender to which the animal belongs'
          type: integer
          format: int64
          example: 1
        color_id:
          description: 'Foreign key identifier for the color to which the animal belongs'
          type: integer
          format: int64
          example: 3
        coat_id:
          description: 'Foreign key identifier for the coat type of the animal'
          type: integer
          format: int64
          example: 4
        sizerange_id:
          description: 'Foreign key identifier for the size range to which the animal belongs'
          type: integer
          format: int64
          example: 5
        agerange_id:
          description: 'Foreign key identifier for the age range to which the animal belongs'
          type: integer
          format: int64
          example: 6
      type: object
...
  securitySchemes:
    bearerAuth:
      type: http
      description: 'JWT Authorization header using the Bearer scheme.'
      bearerFormat: JWT
      scheme: bearer